+++ bindings/nostr-sdk-js/src/client/mod.rs
@@ -28,7 +28,7 @@ use self::zapper::{JsZapDetails, JsZapEntity};
 use crate::abortable::JsAbortHandle;
 use crate::database::JsNostrDatabase;
 use crate::duration::JsDuration;
-use crate::pool::result::{JsSendEventOutput, JsSendOutput};
+use crate::pool::result::{JsReconciliationOutput, JsSendEventOutput, JsSendOutput};
 use crate::relay::blacklist::JsRelayBlacklist;
 use crate::relay::options::{JsNegentropyOptions, JsSubscribeAutoCloseOptions};
 use crate::relay::{JsRelay, JsRelayArray};
@@ -735,11 +735,16 @@ impl JsClient {
     /// Negentropy reconciliation
     ///
     /// <https://github.com/hoytech/negentropy>
-    pub async fn reconcile(&self, filter: &JsFilter, opts: &JsNegentropyOptions) -> Result<()> {
+    pub async fn reconcile(
+        &self,
+        filter: &JsFilter,
+        opts: &JsNegentropyOptions,
+    ) -> Result<JsReconciliationOutput> {
         self.inner
             .reconcile(filter.deref().clone(), **opts)
             .await
             .map_err(into_err)
+            .map(|o| o.into())
     }
 
     /// Handle notifications