+++ crates/nostr-relay-pool/src/pool/error.rs
@@ -43,6 +43,9 @@ pub enum Error {
     /// Event/s not published
     #[error("event/s not published")]
     EventNotPublished,
+    /// Negentropy reconciliation failed
+    #[error("negentropy reconciliation failed")]
+    NegentropyReconciliationFailed,
     /// Relay not found
     #[error("relay not found")]
     RelayNotFound,