+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 * nostr: add `EventBuilder::interest_set` ([Yuki Kishimoto])
 * nostr: add `title`, `image` and `description` constructors to `Tag` ([Yuki Kishimoto])
 * pool: add `SendOutput` and `SendEventOutput` structs ([Yuki Kishimoto])
+* pool: return `ReconciliationOutput` for all `reconcile*` methods ([Yuki Kishimoto])
 * signer: add `NostrSigner::unwrap_gift_wrap` method ([Yuki Kishimoto])
 * sdk: add `Client::unwrap_gift_wrap` method ([Yuki Kishimoto])
 * js(sdk): partially expose `JsRelayPool` ([Yuki Kishimoto])